Reorganize package under smart_price namespace
