+++ smart_price/core/common_utils.py
