+++ smart_price/extract_excel.py
@@ -1,0 +1,2 @@
+"""Wrapper module for convenience imports."""
+from smart_price.core.extract_excel import *  # noqa: F401,F403
