+++ tests/test_extract_pdf.py
@@ -2,7 +2,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-import extract_pdf
+from smart_price import extract_pdf
 
 
 def test_esmaksn_pdf_threshold():
